Add GA submission grades handling; implement batch creation and replacement for rubric grades

+++ Server/src/main/java/com/grade/forge/grading/dto/SubmissionGradeItemRequest.java
@@ -13,7 +13,7 @@ import lombok.Setter;
 @Builder
 public class SubmissionGradeItemRequest {
     private Long rubricSubCriteriaId;
-    private Integer awardedScore;
+    private Double awardedScore;
     private String feedback;
 }
