+++ Server/src/main/java/com/grade/forge/grading/dto/SubmissionGradeItemResponse.java
@@ -13,7 +13,7 @@ import lombok.Setter;
 @Builder
 public class SubmissionGradeItemResponse {
     private Long rubricSubCriteriaId;
-    private Integer awardedScore;
+    private Double awardedScore;
     private String feedback;
 }
