+++ Server/src/main/java/com/grade/forge/grading/dto/SubmissionGradeResponse.java
@@ -18,7 +18,7 @@ public class SubmissionGradeResponse {
     private String rubricSubCriteriaDescription;
     private Long rubricCriteriaId;
     private String rubricCriteriaTitle;
-    private Integer awardedScore;
+    private Double awardedScore;
     private String feedback;
 }
